requirements.txt
fastapi
uvicorn[standard]
python-dotenv
playwright
httpx
beautifulsoup4
pandas
PyPDF2
openai
openai-whisper
faster-whisper
soundfile
